refactor(admin): decompose AdminOperationsController behind jobs/overview/logs facades (P10.1)
Decompose the 1677-line AdminOperationsController into a thin controller
(validate request -> call facade -> record audit event -> map to IActionResult)
backed by three new facade services in Transcendence.Service.Core:

- IAdminJobsFacade / AdminJobsFacade: all jobs/* logic (recurring list +
  trigger/pause/resume, queues, list, failed, inspect/{id}, failed/{id}, retry,
  inspect/{id}/delete, bulk-delete). Owns JobStorage, IRecurringJobManager,
  IWorkerRecurringJobPolicy, WorkerJobScheduleOptions, MultiRegionIngestionOptions
  and the jobs-only consts/sets. Returns result/DTO records so the controller
  keeps HTTP-status mapping and audit recording.
- IAdminOverviewFacade / AdminOverviewFacade: overview + metrics/analysis. Owns
  the EF context, ingestion/multi-region options and JobStorage for the backlog
  snapshot.
- IAdminLogsFacade / AdminLogsFacade: logs/services tail. Owns IConfiguration +
  the allowed service-log keys.

The admin DTO contracts moved into Service.Core but stay in the
Transcendence.WebAPI.Controllers namespace so the public OpenAPI schema names and
the controller call sites are unchanged. Facades resolve Hangfire's JobStorage
(Hangfire.Core, already referenced by Service.Core); registered via a new
AddAdminOperationsFacades() called from the WebAPI host (kept out of
AddTranscendenceCore to preserve that method's Hangfire-free DI-validation
contract). cache/invalidate and audit-log stay thin (direct service calls);
mutating-op audit recording stays cross-cutting in the controller.

Behavior-preserving: routes, request validation, response shapes/DTOs, status
codes, NotFound/BadRequest paths and audit-event metadata are byte-for-byte
equivalent. Gated by AdminOperationsControllerTests (now constructing the real
facades); closes the P10.1 admin half. Controller: 1677 -> 369 lines.
